renamed Profile components, removed Button condition
