+++ src/pages/Profile/settings/EditInfo.tsx
@@ -1,33 +1,35 @@
 import React, { FC } from 'react';
 
 import { UserData } from '../../../api/UserAPI';
+import { Button } from '../../../components/ui/Button/Button';
 import { MainLink } from '../../../components/ui/Link/Link';
 import { TRANSLATION } from '../../../lang/ru/translation';
 import { InformationForm } from '../InformationForm';
 import { UserThumbnail } from '../UserThumbnail';
 
-const { LinkToBack, LinkToChangeInfo, LinkToChangePassword } = TRANSLATION.Profile;
+const { LinkToBack, SaveButton } = TRANSLATION.Profile;
 
-interface MainProps {
-  url: string;
+interface EditProps {
   userInfo: UserData;
+  url: string;
 }
 
-export const Main: FC<MainProps> = ({ url, userInfo }) => {
+export const EditInfo: FC<EditProps> = ({ url, userInfo }) => {
   return (
     <>
       <div className="profile__menu">
         <UserThumbnail avatar={userInfo.avatar} login={userInfo.login} />
-        <MainLink text={LinkToChangeInfo} href={`${url}/edit`} extendClass="ui-link--button mb-6" />
-        <MainLink
-          text={LinkToChangePassword}
-          href={`${url}/password`}
-          extendClass="ui-link--button mb-4"
+        <Button
+          form="profile-edit"
+          type="submit"
+          text={SaveButton}
+          typeAction="success"
+          extendClass="mb-4"
         />
-        <MainLink text={LinkToBack} href="/" />
+        <MainLink text={LinkToBack} href={`${url}`} />
       </div>
       <div className="profile__inputs">
-        <InformationForm disabled={true} userInfo={userInfo} />
+        <InformationForm disabled={false} userInfo={userInfo} />
       </div>
     </>
   );
