+++ src/pages/Profile/settings/Settings.tsx
@@ -1,35 +1,33 @@
 import React, { FC } from 'react';
 
 import { UserData } from '../../../api/UserAPI';
-import { Button } from '../../../components/ui/Button/Button';
 import { MainLink } from '../../../components/ui/Link/Link';
 import { TRANSLATION } from '../../../lang/ru/translation';
 import { InformationForm } from '../InformationForm';
 import { UserThumbnail } from '../UserThumbnail';
 
-const { LinkToBack, SaveButton } = TRANSLATION.Profile;
+const { LinkToBack, LinkToChangeInfo, LinkToChangePassword } = TRANSLATION.Profile;
 
-interface EditProps {
-  userInfo: UserData;
+interface MainProps {
   url: string;
+  userInfo: UserData;
 }
 
-export const ModifyData: FC<EditProps> = ({ url, userInfo }) => {
+export const Settings: FC<MainProps> = ({ url, userInfo }) => {
   return (
     <>
       <div className="profile__menu">
         <UserThumbnail avatar={userInfo.avatar} login={userInfo.login} />
-        <Button
-          form="profile-edit"
-          type="submit"
-          text={SaveButton}
-          typeAction="success"
-          extendClass="mb-4"
+        <MainLink text={LinkToChangeInfo} href={`${url}/edit`} extendClass="ui-link--button mb-6" />
+        <MainLink
+          text={LinkToChangePassword}
+          href={`${url}/password`}
+          extendClass="ui-link--button mb-4"
         />
-        <MainLink text={LinkToBack} href={`${url}`} />
+        <MainLink text={LinkToBack} href="/" />
       </div>
       <div className="profile__inputs">
-        <InformationForm disabled={false} userInfo={userInfo} />
+        <InformationForm disabled={true} userInfo={userInfo} />
       </div>
     </>
   );
